* Removed SQL Lite for SQLCE (Fixed multi-threading crashes)   * All processing and plug-in manager moved to zvsProcessor dll project.  (GUI does not have to run for program to work now)   * All plug-ins run in their own threads.   * All Managers run in their own threads.   * Removed WinForms project...Recreating GUI in WPF. This will be a considerable effort to finish

+++ zVirtualScenesCommon/Entity/builtin_command_que.cs
@@ -16,9 +16,9 @@ namespace zVirtualScenesCommon.Entity
         /// Called when a Builtin command is added to the que
         /// </summary>
         public static event BuiltinCommandAddedEventHandler BuiltinCommandAddedToQueEvent;
-        public delegate void BuiltinCommandAddedEventHandler(long builtin_command_que_id);
+        public delegate void BuiltinCommandAddedEventHandler(int builtin_command_que_id);
 
-        public static void BuiltinCommandAddedToQue(long id)
+        public static void BuiltinCommandAddedToQue(int id)
         {
             if (BuiltinCommandAddedToQueEvent != null)
                 BuiltinCommandAddedToQueEvent(id);
