+++ zVirtualScenesCommon/Entity/device_type_commands.cs
@@ -8,7 +8,7 @@ namespace zVirtualScenesCommon.Entity
 {
     public partial class device_type_commands : EntityObject
     {
-        public void Run(long device_id, string argument = "")
+        public void Run(int device_id, string argument = "")
         {
             device_type_command_que cmd = new device_type_command_que { device_type_command_id = this.id, arg = argument, device_id = device_id };
             cmd.Run();
